fix: password hash
